+++ src/pages/posts/[...slug]/index.png.ts
@@ -28,7 +28,7 @@ export const GET: APIRoute = async ({ props }) => {
   }
 
   const buffer = await generateOgImageForPost(props as CollectionEntry<"blog">);
-  return new Response(buffer, {
+  return new Response(buffer as Uint8Array, {
     headers: { "Content-Type": "image/png" },
   });
 };
